tec: Rename table `user` in `users`
`user` is a reserved keyword in PostgreSQL. When I executed `select *
from user;`, it returned the list of system users instead of the users
of the application. I had to execute `select * from public.user;`.

This could lead to a bad experience for adminsys trying to display the
content of the user table, so I decided to rename the table.

In general, I'm in favor of using plural for tables names, but I didn't
pay attention to the name generated by Symfony.

Reference: https://www.postgresql.org/docs/current/sql-keywords-appendix.html