tec: Fix the setup of Vite
I need to set `assetsDir` to `assets` in order to generate valid URLs
for files called with CSS or JavaScript (e.g. `url()` in `background-image`).

This means that I also need to set `outDir` to `public/`… which contains
the `index.php` file. Unfortunately, I don't want it to be removed when
building, so I *also* need to set `emptyOutDir` to `false`. However, I
still need to empty the `assets/` dir, and there is no options for this
in Vite. So in the end, I had to write a custom plugin to empty the
assets directory.

Reference: https://github.com/vitejs/vite/issues/3541