fix: Fix the CreateTicket migration
I changed the `created_at` and `created_by` columns names while it
wasn't necessary. Worse: I forget to change the index names for MySQL
and so the migrations wasn't synced with the schema.

It's ok to fix the migrations while Bileto isn't released, but later
I'll have to create new migrations for this kind of changes.