fix: Fix "new ticket" form not submitted
It seems that TinyMCE doesn't support textareas with a `required`
attribute. My tests have shown that submitting the form a first time
worked, but not anymore if I was going back to it (probably a side
effect of Turbo).

There’s an old (closed) ticket about this issue on GitHub and the given
workaround works... but I don't understand it so I prefer to simply
remove the attribute. It's ok since TinyMCE doesn't handle it at all
(i.e. even with the attribute, the browser still submits the form).

Reference: https://github.com/tinymce/tinymce/issues/2584