tec: Render the full ticket page when submitting messages
In the next commit, we'll be able to change the status of the ticket
when answering. Since the status appears at the top of the page, it
needs to be rendered in the response.

This brings some difficulties because I no longer use `<turbo-frame>`
and now the scroll is set to the top of the page at the end of the
submit. Also, there are two cases to handle:

1. the data are valid and we redirect to the "show" page: in this case,
   a new `data-turbo-preserve-scroll` attribute allows to cancel the
   scroll of the page
2. an error in the form (e.g. empty content) is detected and we render a
   bad request response: in this case, the `data-turbo-preserve-scroll`
   solution doesn't work (i.e. the Turbo `currentVisit` doesn't exist),
   so we generate an `autofocus` attribute in order to scroll to the
   form.

The user experience is somewhat degraded, but I think it's good enough,
at least for a MVP.