fix: Fix popup not opening after back navigation
The `data-turbo-cache` removes the element from the DOM in the cache.

When we use the "back" button of the browser, Turbo loads the content
from the cache, but doesn't re-fetch the content. It means that the
popup elements is simply not there anymore.

This attribute was used to make sure to hide this element when we go
back, but I'll have to find another solution.