tec: Add more Controller tests
Some tests were missing because `$client->submitForm` cannot set any
kind of value (e.g. it cannot set a value which is not part of the
available options of a select box). I initially used this method
(instead of directly calling `$client->request`) because I wasn't able
to set the CSRF token manually.

Following c3f412e and the introduction of the `SessionHelper`, I finally
found a way to generate this token by getting the session with its id
and setting the token manually.

I've changed all the tests to use this new system, and I added the
missing tests. Along the way, I've detected some bugs that I've fixed.