Always translate errors in controllers
Well, this is a long story.

I noticed that the errors coming from the Entities violations were
already translated. Thus, in the templates, the errors were displayed
as-is (i.e. without the `trans` filter).

However, the errors coming from the controllers had to be translated.
This resulted into an inconsistency between errors because these had to
be `trans`lated. Thus, I started to translate the errors in the
controllers.

Unfortunately, I wasn't able to translate the `csrf.invalid` key because
I was using a method in the `BaseController` (i.e. `csrfError()`) which
had no access to the `Translator` service (injection was not an option,
and the service is not available in the `$this->container`).

So, to avoid another inconsistency, I took out the generation of the
CSRF error. This result into a long list of small changes.