doc: Fix the doc to retrieve the latest Git tag
The previous command retrieved the latest tag from the *current* branch.
It means that during the update, since the repository was configured on
the previous tag, it was returning this one instead of the "real" latest
tag.

This new command returns the latest tag accross all the branches.

Reference: https://stackoverflow.com/a/7979255