tec: Disable XDebug when running PHPStan
As stated in the documentation:

> PHPStan turns off XDebug if it’s enabled to achieve better performance.
> If you need to debug PHPStan itself or your custom extensions and want
> to run PHPStan with XDebug enabled, pass this option.

So we don't want XDebug to be enabled or it degrades the performance.

I've initially added this option because there is a warning telling that
XDebug is active but "--xdebug" is missing and that can slow down
performance. I find this message misleading because adding this option
slows down performance a LOT more.

Reference: https://phpstan.org/user-guide/command-line-usage#--xdebug