tec: Remove UniqueEntity constraint from uid fields
I completely removed the UniqueEntity constraints on uid fields because
this field is set just before the insertion in the database. Meaning it
is set *after* calling validate().

The UidGenerator already makes sure to generate a unique value, even if
there is a very small chance that two parallel requests generate the
same value. I think the risk is acceptable.