doc: Extract documentation indexes in dedicated files
The developers' guide was beginning to be too big to stay in the main
README, so I moved it in a dedicated file. The administrators' guide is
smaller, but it's more coherent to move it as well.

I also moved the maintainers/release.md file under developers/ as it was
the only file and as the maintainers are also developers.