imp: Use the current date instead of the email date
The datetime of the mail server may be different from that of Bileto.
This means that the answers may be out of order, whether they were added
by from emails or from the interface.

This could be an option when creating a Mailbox, but for now we’ll just
use the current date from Bileto.