Allow to start an OpenLDAP server in dev mode
The server is configured as an optional service in the Docker Compose
file. It is started by enabling the `ldap` profile. This can be done
with `make docker-start LDAP=true`.

The users are defined in a ldif file placed under `docker/ldap-ldifs/`
which is mounted as a volume in the Docker Compose file. Only Charlie
and Dominique users are defined as the other users will continue to
login as "local" users.

I've enabled `BITNAMI_DEBUG` to help to understand the errors when we
update the ldif file. It's not very verbose anyway.