Configure a custom FormLoginAuthenticator
The Authenticator is responsible for verifying user credentials.

Symfony provides several Authenticators, including the “Form Login” that
we already use, and a LDAP variant. Unfortunately, there is no way to
log in from a database or a LDAP directory depending on an attribute
(e.g. auth_type). So we had to write our own Authenticator.