imp: Redirect if user can create ticket in its default organization
When the user can create tickets in multiple organizations, the main
"new ticket" shows a form to select the organization.
It is annoying for the users to have to choose.

This patch considers that the default organization is the one in which
the user must create its ticket.
Otherwise, it is still possible to create ticket in a different
organization by going on its page explicitely.