dev: Refactor handling of MessageDocuments
There are few things that I fixed in order to have better code.

First, the tinymce_controller was able to communicate to the
message_documents_controller. This was because tinymce_controller was
included in the other controller. This didn't make sense at all though.
In this patch, tinymce_controller and message_documents_controller are
now only applied on their limited DOM sections. They can communicate via
events handled and transferred by more global controllers
(form_new_ticket_controller and form_new_answer_controller).

Also, I wasn't happy with the controllers manipuling the DOM like they
did. It was error-prone. Now, the MessageDocuments are rendered via a
`<turbo-frame>` which is reloaded when a document is added or removed.
It's easier to handle, while there is a small delay before the documents
being rendered.

An impact of the previous change is that all the unattached
MessageDocuments are attached to the created message. This means that I
no longer have to maintain a list of hidden inputs to pass the list of
MessageDocuments uids to the server. This also means that I will not
have to clean the unattached documents as they appear clearly to the
user.

These changes allowed me to clean some parts:

- removing the images in the editor is now handled directly by the
  tinymce_controller
- I was able to remove some (now) useless fields from the
  MessageDocumentsController#create endpoint
- I removed the templated message document element