fix: Fix size of images in messages
The users can resize their images in messages. However, only the height
was kept by the HTML sanitizer. The height is set to `auto` in CSS
though. The consequence was that none of width or height were kept in
the rendered message.

I need to keep `height: auto` because `max-width` is set to 100% in
order to keep the sizes ratio. The only remaining solution is to allow
the `width` attribute on `img` in the HTML sanitizer which is fine.

This should be fine as the default W3C reference states that in their
default configuration, `width` is allowed on all elements (while the
HTMLSanitizer component disallows it).

References:

- https://wicg.github.io/sanitizer-api/#default-configuration
- https://github.com/symfony/symfony/blob/6.3/src/Symfony/Component/HtmlSanitizer/Reference/W3CReference.php