dev: Remove rollbacks test
As explained in https://github.com/Probesys/bileto/issues/230:

I tried to rewrite the tests about the migrations. I was able to test
that no new migration was to be generated, but I realized that testing
the rollbacks wasn't working as expected. Indeed, the current migrations
version is already set to `0` (i.e. no migration applied). By
reorganizing the test and making sure to apply the migrations before the
rollbacks, it doesn't work better because of how Doctrine works. Indeed,
in the Doctrine/Migrations/Version/DbalExecutor class, Doctrine
retrieves the SQL queries to run for each migration. Unfortunately,
these queries are never reset. So the queries from the initial
`doctrine:migrations:migrate` command are still present in memory. It
means that when running the rollback command, Doctrine will first try to
execute the `up()` queries of the last migration. It fails because the
column/table/whatever already exists in the database. A solution would
be to execute this test outside of PHP.