dev: Improve signature of CommandTestsHelper::executeCommand
We need more often to pass arguments than passing direct inputs, so I
switched the order of both arguments (and I fixed type hinting).

I've also set "interactive" mode to false when the inputs are not given
to avoid warning in commands that support the `--no-interaction` option
(passing it as argument doesn't work).