dev: Refactor the transfer of events of the editor
There is a transfer of events between the TinyMCE and the
MessageDocuments JS controllers. Before, it was handled with the same
code duplicated in two different controllers.

Now, there is a single `editor` controller to handle that.
The FormNewTicket controller is removed as it was used only for this
purpose.