fix: Fix the year format in the Twig dateTrans filter
In ICU, `Y` is the year of “Week of Year”. It means that if the 1st
January is part of week 52 of the previous year, the displayed year will
be the previous year.

`y` and `yyyy` are the same, but I find it more clear that `yyyy` is
different than `Y`. It's also more coherent with `d` being the day on
one character and `dd` being the day on two characters.

Reference: https://unicode-org.github.io/icu/userguide/format_parse/datetime/