fix: Don't track empty entity changes
Doctrine doesn't trigger the postUpdate event for the ManyToMany changes.
However, this kind of change set the `updatedBy` and `updatedAt` fields.
So postUpdate is triggered for these fields.
However, we don't want to track these changes (see the comment in the
EntityActivitySubscriber class) and so the returned `entityChanges` is
empty.

Before, we were creating an EntityEvent in all cases.
Now, it's only created if `entityChanges` is not empty.