Rebalance the font size
This allows to put more text on the screen and make the text less
agressive.

In some places, I've removed the "small" text as it looks ok with normal
size. This decreases the visual inconsistencies and it's less tiring on
the eyes.

I've extracted variables so it will be easier to adapt the size of the
text later.