dev: Restore security files on make translations
For some reasons, Symfony doesn't detect the security keys and remove
them from the security files. With this patch, we restore these files
immediately with Git.

The downside is that we'll not be able to use `make translations` to
update the security files. Hopefully, we don't change them often (if
ever).