fix: Track changes to the tickets' ongoing contracts
Doctrine doesn't dispatch the `postUpdate` event on ManyToMany relations
changes. It means that we cannot track these changes in the
EntityActivitySubscriber.

It seems to be doable with the `onFlush` event, but I wasn't succeeded
without spending too much time on this subject.

Reference: https://stackoverflow.com/q/31616315

Instead, I preferred to log the change manually, even though it's a bit of a
hack. I hope to find a better solution in the future, but it's good
enough for now.