dev: Make deprecations notices less verbose in tests
There are a lot of indirect deprecations notices that I can't fix for
the moment. They take a lot of space and the results of the tests are
harder to read. So I'm making them less visible with the risk of missing
some important deprecations that I could fix. It should be a temporary
fix (but we all know that "temporary" means "for an undefined amount of
time" :)).