dev: Add getEntityType to RecordableEntityInterface
In some cases, the entity passed to the EntityEvent was a Doctrine
proxy. When calling `$entity::class`, the proxy class was returned
instead of the real class.

In this solution, `self::class` returns the expected class. Note that
both `static::class` or `get_class($entity)` would return the proxy
class.