imp: Improve the lifecycle of the tickets
I had to refactor the code related to the lifecycle of the tickets
before adding the "accept/refuse solution" feature. Unfortunately, I've
changed quite a lot of things along the way so this commit may look like
difficult to read.

In short, I've extracted the business logic related to some ticket's
events (e.g. ticket is created, an answer is posted). Subscribers listen
for these events to perform actions on the ticket:

- MessageDocumentsSubscriber attaches MessageDocument to the posted
  message
- OngoingContractSubscriber attaches an ongoing contract to the created
  ticket
- LifecycleSubscriber handles all the logic of changing the ticket's
  status based on the event

The lifecycle of the tickets changed a bit:

- when assigned, a new ticket changes to "in progress"
- when an assignee answers to an "in progress" ticket, it changes to
  "pending"
- when a requester answers to a "pending" ticket, it changes to "in
  progress"
- when an assignee answers with a solution, the ticket changes to
  "resolved"
- when a requester answers to a resolved ticket, it changes to "in
  progress"

In consequence, I had to change the form to answer to a ticket.
Fortunately, it allows to have the same behaviours when the answers come
from the application or from an email.