misc: Migrate the entity_event changes for MariaDB
I don't understand this migration for MariaDB, but I guess it's
important. That's something that I dislike with this kind of tool: weird
things can happen, not letting you understand easily what's going on.

The issue doesn't happen with PostgreSQL. I think it was a bug in
Doctrine that has been fixed during the last update of the dependencies.