fix: Ignore HTML errors when creating tickets from emails
I had an error because an email contained a tag `<o:p>` (the `o`
namespace wasn't defined). The fetching was failing quite badly because
of this error.

I think we should just ignore this kind of errors as they are produced
by bad HTML. It should be ok since we sanitize the HTML later.