dev: Allow to check permissions of any user
This commit introduces the class App\Security\Authorizer to check
permissions within the application. It replaces the standard Security
class when calling `isGranted`, and it adds the `isGrantedToUser`
method.

I could have continued to use Security when calling `isGranted`, but it made
the calls less consistent within the application.
I couldn't extend the Security class as it is final.