fix: Load only "orga" authorizations when checking for "any"
When we checked an `orga:*` permission with a scope of "any", we also
loaded the admin authorizations. This wasn't a security problem, as the
admin and orga permissions are separated in different roles, but it was
a potential risk anyway.

I refactored the code so that we use the same code to load the
authorizations when the subject is "any" or an Organization. This should
also improve the performance: we now load the role in the mean time as
the organization when the subject is "any".