dev: Output PHPUnit results with TestDox format
I removed the TestDox output when upgrading to PHPUnit 10 because the
behaviour changed: TestDox is output at the end, once all the tests have
run. I found this useless and so I removed this output.

However, I now realize that it can help to understand the order in which
the tests are run. It can help to debug strange behaviour.

Reference: 4dd097d