dev: Move TicketRepository methods related to TicketSearcher
It was difficult to remember that TicketRepository had to be adapted
when changing code in the TicketQueryBuilder. I wasn't happy with this
design where some code related to the TicketSearcher wasn't under the
SearchEngine/ folder.

Now, the full creation of the Doctrine QueryBuilder is delegated to
TicketQueryBuilder, and its manipulation is delegated to TicketSearcher.