Allow to filter tickets by labels
The label filter is special as its different values are searched with a
"AND" condition. Said otherwise, when a user select several labels to
filter, we search for tickets that have all the selected labels instead
of tickets that have one of the selected labels. This is different from
the other filters.

The consequence is that this filter is handled a bit differently by the
TicketFilter. The implementation is not perfect, but I don't want to
spend too much time on this specific case.