fix: Fix contracts pagination
When importing production data, I discovered that the pagination of the
contracts was broken. In some complex cases, the contracts are not
correctly fetched from the database. It involves contracts with tickets
and/or spent time. The `JOIN`s seem to be the problem.

As I use Doctrine Pagination for this, I searched for a solution. The
documentation mentions the `fetchJoinCollection` argument to handle
queries with many-to-many joins. Unfortunately, it doesn't work to fix
the issue.

Since I can't find a solution, I just remove the `JOIN`. This means a
lot of N+1 requests and worse performance. But at least the data is
correct. Help is welcome to improve the performance while keeping the
pagination correct :)