fix: Make sure to return valid UTF8 string in DomHelper
I wasn't able to create a string that would produce the issue, but in
some cases, $dom->saveHTML() doesn't return a valid UTF8 string.

The consequence was that the HTML sanitizer would return an empty
string.

Strangely, `saveHTML()` returns a valid UTF8 string when passing the
`$dom->documentElement`.

Reference: https://stackoverflow.com/a/20675396