fix: Keep current actors in the ticket after losing access
When a user was requester or observer of a ticket, but was losing access
to the corresponding organization, the user was no longer part of the
"actors" lists of the ticket.

It means that when an agent was then changing the person assigned to
such a ticket, the user was removed from the ticket and the agent was
set as the requester. This was a really weird behaviour for agents.

Now, we consider that, as the user has actually access to the ticket, it
can be safely returned by the ActorsLister.