imp: Allow to select the organization when opening a ticket from the global list
Before, clicking on "New ticket" redirected to the "new ticket" form of
the current user's default organization. However, agents often need to
open a ticket in a different organization. In this case, they needed to go
to the organizations' list, find the organization, open it, and click on
"new ticket".

Also, the behavior was inconsistent with the one of the "contracts"
list.

Now, if the user can create a ticket in several organizations, a modal
is opened to allow the user to select the organization.