imp: Reopen resolved ticket if answerer is not an agent
There are cases where an observer or a user allowed to access all the
tickets answers to a "resolved" ticket. In this case, the ticket stayed
in this status. However, in many cases, the ticket needs the attention
of the agents.

Now, the ticket is reopened and the solution removed, so the agents
cannot miss a "not so resolved" ticket.