imp: Change quick search so it searches for tickets with any label
All the filters of the quick search form are acting as "OR" (selecting
several statuses or priorities lists tickets that have status1 OR
status2 / priority1 OR priority2). However, the labels behaved
differently (label1 AND label2). This different behaviour was quite
disturbing for new users.

Now, the behaviour of labels is coherent with the other fields.