fix: Allow invited users to answer if they have no organization
A user who had only the `orga:create:tickets:messages` permission and no
default organization wouldn't be able to answer to a ticket in which he
was observer. It was an annoying issue as it forced us to create a fake
organization in which they had the `orga:create:tickets` permission,
just so they were able to answer to tickets in other organizations.

I took the opportunity of this fix to reorganize the file a bit in order
to identify more easily the important parts of the processMailboxEmail
method.