fix: Fix PHP Docker images failing because of missing libc-client-dev
Image php:8.2 is now based on Debian Trixie by default. Trixie no longer
provides the `libc-client-dev` package, which is required for the PHP
`imap` extension.

Also, this extension will no longer be packaged with PHP 8.4+. It means
it's not a good solution for the future.

I don't see any solution other than:

1. base the images on Bookworm (it should be maintained until the next
   major version of Debian);
2. then later, remove the `imap` extension requirement, which could
   require to fix an issue in the `webklex/php-imap` dependency.

References:

- Repository of the (unmaintained) `imap` extension:
  https://github.com/php/pecl-mail-imap
- Issue about the missing package in the PHP image:
  https://github.com/docker-library/php/issues/1608
- The PR adding the `imap` requirement:
  https://github.com/Probesys/bileto/pull/510