fix: Fix sending email notifications
Previously, getting the Message-ID of sent emails wasn't reliable and
sometimes returned ids not compliant with RFC 2822. As these ids are
used then in the References and In-Reply-To headers of the email
notifications concerning the future answers, Symfony failed badly with
an exception.

Now, the logic is improved to get the correct Message-ID of the sent
emails. We also set References and In-Reply-To headers only with values
compliant with the RFC.